Refines Model and DTO for integer.

+++ src/main/java/de/superrental/carrentalinterface/dto/CarDTO.java
@@ -4,5 +4,5 @@ public record CarDTO(
         String brand,
         String model,
         Boolean inRental,
-        Long carId
+        Integer carId
 ) {}
